Finished just need comments
Finished with the entire project; all that is left are comments for the index page and the css page. The last few steps in the project that I finished were just styles. I created a style that will affect the way the video and media player will look on devices of a certain width. I made it so that the media player would be 360 pixels wide and will be displayed on the right. I also added a style that will be applicable to all devices; it will affect the look of the captions by changing their color, background color, and giving them a slightly larger font size.